+++ src/wdio/services/SeleniumDockerService.js
@@ -58,6 +58,7 @@ export default class SeleniumDockerService {
       image: null, // The image name to use, defaults to selenium/standalone-${browser}
       retries: 500, // Retry count to test for selenium being up
       retryInterval: 10, // Retry interval in milliseconds to wait between retries for selenium to come up.
+      env: {}, // Environment variables to add to the container
       ...(config.seleniumDocker || {}),
     };
     this.host = config.host;
@@ -69,8 +70,20 @@ export default class SeleniumDockerService {
     return new Promise((resolve, reject) => {
       if (this.config.enabled) {
         const containerId = this.getContainerId();
+
+        // Workaround for browser specific docker issues
+        // see https://github.com/SeleniumHQ/docker-selenium#running-the-images
+        let args = '';
+        if (this.browserName === 'chrome') {
+          args = '-v /dev/shm:/dev/shm';
+        } else if (this.browserName === 'firefox') {
+          args = '--shm-size 2g';
+        }
+
+        const env = Object.keys(this.config.env).map(key => `-e ${key}=${this.config.env[key]}`).join(' ');
+
         if (!containerId) {
-          exec(`docker run --rm --cidfile ${this.cidfile} -p ${config.port}:4444 ${this.getImage()}`);
+          exec(`docker run -d --rm ${env} ${args} --cidfile ${this.cidfile} -p ${config.port}:4444 ${this.getImage()}`);
         }
         // Retry for 500 times up to 5 seconds for selenium to start
         retry({ times: this.config.retries, interval: this.config.retryInterval },